Changes to enable patching with omitted directories
Not a C# developer in the slightest (working knowledge of C++). There may be bugs. This update solves an issue when mods are shard through services like Github that strip the empty directories from uploads. The update allows one to patch a game by checking for each directory and creating an empty one if it isn't found.